Extract s3 uploading/deleting into an enum
So that next we can make a variant that handles local upload/delete.
