+++ src/app.rs
@@ -7,10 +7,9 @@ use conduit_middleware::Middleware;
 use git2;
 use oauth2;
 use r2d2;
-use s3;
 use curl::easy::Easy;
 
-use {db, Config};
+use {db, Config, Uploader};
 
 /// The `App` struct holds the main components of the application like
 /// the database connection pool and configurations
@@ -23,7 +22,7 @@ pub struct App {
 
     /// The GitHub OAuth2 configuration
     pub github: oauth2::Config,
-    pub bucket: s3::Bucket,
+    pub uploader: Uploader,
     pub s3_proxy: Option<String>,
     pub session_key: String,
     pub git_repo: Mutex<git2::Repository>,
@@ -64,11 +63,7 @@ impl App {
             database: db::pool(&config.db_url, db_config),
             diesel_database: db::diesel_pool(&config.db_url, diesel_db_config),
             github: github,
-            bucket: s3::Bucket::new(config.s3_bucket.clone(),
-                                    config.s3_region.clone(),
-                                    config.s3_access_key.clone(),
-                                    config.s3_secret_key.clone(),
-                                    config.api_protocol()),
+            uploader: Uploader::new_s3(&config),
             s3_proxy: config.s3_proxy.clone(),
             session_key: config.session_key.clone(),
             git_repo: Mutex::new(repo),
