+++ src/util/io_util.rs
@@ -1,5 +1,6 @@
 use std::io::prelude::*;
 use std::io;
+use std::mem;
 
 pub struct LimitErrorReader<R> {
     inner: io::Take<R>,
@@ -21,4 +22,26 @@ impl<R: Read> Read for LimitErrorReader<R> {
             e => e,
         }
     }
+}
+
+pub fn read_le_u32<R: Read + ?Sized>(r: &mut R) -> io::Result<u32> {
+    let mut b = [0; 4];
+    read_fill(r, &mut b)?;
+    Ok(((b[0] as u32) <<  0) |
+       ((b[1] as u32) <<  8) |
+       ((b[2] as u32) << 16) |
+       ((b[3] as u32) << 24))
+}
+
+pub fn read_fill<R: Read + ?Sized>(r: &mut R, mut slice: &mut [u8])
+                                   -> io::Result<()> {
+    while slice.len() > 0 {
+        let n = r.read(slice)?;
+        if n == 0 {
+            return Err(io::Error::new(io::ErrorKind::Other,
+                                      "end of file reached"))
+        }
+        slice = &mut mem::replace(&mut slice, &mut [])[n..];
+    }
+    Ok(())
 }
